+++ Service/MessageService.cs
@@ -10,8 +10,7 @@ using System.Threading.Tasks;
 namespace BoTools.Service
 {
     public class MessageService
-    {
-        private static string _readTheRulesRole = "🥉";
+    {        
         private static string _eternalInvite = "https://discord.gg/g43kWat";
         #region emote                
         private static readonly string _coinEmote = "<a:Coin:637802593413758978>";
@@ -28,12 +27,7 @@ namespace BoTools.Service
         #region emoji
         private static readonly string _coeurEmoji = "\u2764";
         private static readonly string _tvEmoji = "\uD83D\uDCFA";
-        #endregion
-        #region Role
-        private static readonly ulong _rulesMsgId = 847145384387411989;
-        private static readonly ulong _rulesChannelId = 846694705177165864;
-        #endregion
-        private IRole _IRoleRules = null;
+        #endregion        
         private DiscordSocketClient _client;
         private ISocketMessageChannel _logChannel;        
         private static readonly ILog log = LogManager.GetLogger(MethodBase.GetCurrentMethod().DeclaringType);
@@ -48,8 +42,7 @@ namespace BoTools.Service
             _client.Ready += Ready;            
             _client.UserLeft += UserLeft; // not working
             _client.UserJoined += UserJoined;
-            _client.InviteCreated += InviteCreated;
-            _client.GuildMembersDownloaded += GuildMembersDownloaded;
+            _client.InviteCreated += InviteCreated;            
         }
 
 
@@ -67,10 +60,6 @@ namespace BoTools.Service
             await _client.DownloadUsersAsync(Helper.GetZderLands(_client)); // DL all user
         }
 
-        private async Task GuildMembersDownloaded(SocketGuild arg)
-        {
-            await CheckRole();
-        }
 
         private async Task UserJoined(SocketGuildUser guildUser)
         {            
@@ -240,51 +229,7 @@ namespace BoTools.Service
                 $"En cas de soucis merci de contacter <@!312317884389130241>");
             return;
         }
-        #endregion
-
-        #region Rôle
-        private async Task CheckRole()
-        {
-            if(_IRoleRules == null) _IRoleRules = Helper.GetRole(_client, _readTheRulesRole);
-            await CheckRules();
-            //await CheckAttribution(); ////////////////////////////////////////////////////////////////////////////
-        }
-
-        private Task CheckAttribution() 
-        {
-            // MAX 20reac / message
-            throw new NotImplementedException();
-        }
-
-        private async Task CheckRules()
-        {
-            var channelRules = Helper.GetSocketMessageChannel(_client, "rules");
-
-            IReadOnlyCollection<IMessage> iMsg = channelRules.GetMessagesAsync(1).FirstAsync().Result;
-            IMessage msg = iMsg.First();
-
-            List<SocketGuildUser> allUsers = Helper.GetZderLand(_client).Users.ToList();
-            allUsers.RemoveAll(x => x.IsBot);
-
-            foreach(var user in allUsers)
-            {                
-                await user.RemoveRoleAsync(_IRoleRules);//purge
-            }
-
-            List<IReadOnlyCollection<IUser>> reactListUsers = msg.GetReactionUsersAsync(msg.Reactions.FirstOrDefault().Key, 1000).ToListAsync().Result;
-
-            foreach (var userLst in reactListUsers)
-            {                    
-                var okUserslist = userLst.ToList();
-
-                foreach (var okUser in okUserslist)
-                {
-                    var subject = allUsers.First(x => x.Username == okUser.Username);
-                    await subject.AddRoleAsync(_IRoleRules); //add
-                }                                                      
-            }                 
-        }
-        #endregion
+        #endregion        
 
         #region Embed
         /// <summary>
@@ -350,10 +295,7 @@ namespace BoTools.Service
         public string GetTvEmoji() { return _tvEmoji; }
         #endregion
 
-        private static bool IsStaffOrBot(IUser user)
-        {            
-            return (user.IsBot || user.Username.StartsWith("Vince"));
-        }
+
         private static bool IsStaffOrBotMsg(SocketMessage msg)
         {
             return (msg.Author.IsBot || msg.Author.Username.Trim().StartsWith("Vince"));
